[Spesh] Tear down spesh worker in full_cleanup
If the spesh worker is still active the GC can try to free a lock that
is being held, which leads to an abort in libuv. Stopping the thread
makes full-cleanup work (although it still leaks).

Thread is stopped by signalling with VMNull. Ideally we'd unshift this
to the front of the queue, but that is more complicated.