Fix JIT of specialized bigint ops
In the case where the result was going into the same register that an
input argument came from, we stored the result prematurely, and then
read a NULLed out bigint struct as one of the arguments. This only
took place in the fallback path and when the earlier requirement was
meant. Fixes various crashing spectests.